backend/requirements.txt
algoliasearch-django>=3.0,<4.0
django>=4.0.0,<4.1.0
djangorestframework
django-environ
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
drf-spectacular
django-filter